+++ backend/clickhouse/migrations/000093_add_secure_id_to_errors.up.sql
@@ -1,0 +1,2 @@
+alter table error_groups
+    add column SecureID String;