+++ frontend/src/pages/Auth/WorkspaceInvitation.tsx
@@ -45,9 +45,9 @@ export const WorkspaceInvitation = () => {
 		(w) => w?.id && w.id === workspaceId,
 	)
 
-	const clearInviteAndRedirect = () => {
+	const clearInviteAndRedirect = (projectId?: string | undefined) => {
 		setInviteCode('')
-		navigate('/')
+		navigate(`/${projectId ?? ''}`)
 	}
 
 	useEffect(() => {
@@ -68,7 +68,7 @@ export const WorkspaceInvitation = () => {
 		}
 
 		if (alreadyInWorkspace) {
-			clearInviteAndRedirect()
+			clearInviteAndRedirect(data?.workspace_for_invite_link.project_id)
 			message.success('You are already a member of this workspace.')
 		}
 		// eslint-disable-next-line react-hooks/exhaustive-deps
@@ -123,7 +123,10 @@ export const WorkspaceInvitation = () => {
 										`Successfully joined workspace "${workspaceName}"!`,
 									)
 
-									clearInviteAndRedirect()
+									clearInviteAndRedirect(
+										data?.workspace_for_invite_link
+											.project_id,
+									)
 								} catch (_e) {
 									message.error(
 										'Failed to join the workspace. Please try again.',