+++ frontend/src/pages/Traces/TracePanel.tsx
@@ -56,10 +56,6 @@ export const TracePanel: React.FC = () => {
 		)
 	}
 
-	useHotkeys('esc', () => {
-		traceDialogStore.hide()
-	})
-
 	useHotkeys(
 		'j',
 		() => {