+++ frontend/src/testConfig.tsx
@@ -1,9 +1,9 @@
 import '@fontsource/poppins'
-import '@highlight-run/rrweb/dist/rrweb.min.css'
 import './index.css'
 import './style/tailwind.css'
 import '/__generated/antd.css'
 import '/__generated/index.css'
+import 'rrweb/dist/rrweb.min.css'
 
 import * as React from 'react'