+++ sdk/client/src/listeners/network-listener/utils/utils.ts
@@ -206,9 +206,6 @@ export const matchPerformanceTimingsWithRequestResponsePair = (
 				performanceTiming.toJSON = function () {
 					return {
 						initiatorType: this.initiatorType,
-						// deprecated, use the absolute versions instead
-						startTime: this.startTime,
-						responseEnd: this.responseEnd,
 						// offset by `window.performance.timeOrigin` to get absolute timestamps
 						startTimeAbs:
 							window.performance.timeOrigin + this.startTime,