+++ sdk/client/src/workers/types.ts
@@ -1,4 +1,4 @@
-import { eventWithTime } from '@highlight-run/rrweb-types'
+import { eventWithTime } from '@rrweb/types'
 import { ConsoleMessage, ErrorMessage } from '../types/shared-types'
 import { MetricCategory } from '../types/client'