+++ sdk/highlight-next/src/server.ts
@@ -4,7 +4,9 @@ import * as withHighlightNodeJsAppRouter from './util/with-highlight-nodejs-app-
 import type { ExtendedExecutionContext, HighlightEnv } from './util/types'
 import type { EdgeHandler } from './util/with-highlight-edge'
 import type { NextFetchEvent, NextRequest } from 'next/server'
+import { isNodeJsRuntime } from './util/is-node-js-runtime'
 
+export { isNodeJsRuntime } from './util/is-node-js-runtime'
 export { registerHighlight } from './util/register-highlight'
 export type { HighlightEnv } from './util/types'
 export { H } from '@highlight-run/node' // Imports from server.edge.ts for the edge runtime
@@ -18,7 +20,7 @@ export const Highlight = PageRouterHighlight
 export function PageRouterHighlight(
 	env: HighlightEnv,
 ): PageRouterHighlightHandler {
-	if (isNodeJs()) {
+	if (isNodeJsRuntime()) {
 		return withHighlightNodeJsPageRouter.Highlight(env)
 	} else {
 		throw new Error(
@@ -28,7 +30,7 @@ export function PageRouterHighlight(
 }
 
 export function AppRouterHighlight(env: HighlightEnv) {
-	if (isNodeJs()) {
+	if (isNodeJsRuntime()) {
 		return withHighlightNodeJsAppRouter.Highlight(env)
 	} else {
 		throw new Error(
@@ -46,11 +48,4 @@ export function EdgeHighlight(
 	event: NextFetchEvent & ExtendedExecutionContext,
 ) => Promise<Response> {
 	throw new Error(`unsupported NEXT_RUNTIME: ${process.env.NEXT_RUNTIME}`)
-}
-
-function isNodeJs() {
-	return (
-		typeof process.env.NEXT_RUNTIME === 'undefined' ||
-		process.env.NEXT_RUNTIME === 'nodejs'
-	)
 }