+++ sdk/highlight-next/src/util/register-highlight.ts
@@ -1,7 +1,8 @@
 import type { NodeOptions } from '@highlight-run/node'
+import { isNodeJsRuntime } from './is-node-js-runtime'
 
 export async function registerHighlight(nodeOptions: NodeOptions) {
-	if (process.env.NEXT_RUNTIME === 'nodejs') {
+	if (isNodeJsRuntime()) {
 		const { H } = await import('@highlight-run/node')
 
 		H.init(nodeOptions)